:truck: chore(api) move `sessions-import.js` to `utils`
