+++ api/src/certification/enrolment/application/session-mass-import-controller.js
@@ -2,7 +2,7 @@ import { getI18nFromRequest } from '../../../shared/infrastructure/i18n/i18n.js'
 import * as csvSerializer from '../../../shared/infrastructure/serializers/csv/csv-serializer.js';
 import * as csvHelpers from '../../shared/application/helpers/csvHelpers.js';
 import { usecases } from '../domain/usecases/index.js';
-import { getCsvHeaders } from '../infrastructure/files/sessions-import.js';
+import { getCsvHeaders } from '../infrastructure/utils/sessions-import.js';
 
 const createSessions = async function (request, h) {
   const authenticatedUserId = request.auth.credentials.userId;
